--- a/aten/src/ATen/Declarations.cwrap
+++ b/aten/src/ATen/Declarations.cwrap
@@ -1497,20 +1497,6 @@
     - THGenerator* generator
 ]]
 [[
-  name: _th_exponential_
-  types:
-    - floating_point
-  backends:
-    - CPU
-  cname: exponential
-  variants: function
-  return: self
-  arguments:
-    - THTensor* self
-    - double lambd
-    - THGenerator* generator
-]]
-[[
   name: _th_geometric_
   backends:
     - CPU
@@ -1522,7 +1508,6 @@
     - double p
     - THGenerator* generator
 ]]
-
 [[
   name: _th_copy_ignoring_overlaps_
   cname: copyIgnoringOverlaps