--- a/aten/src/ATen/native/cpu/UnaryOpsKernel.cpp
+++ b/aten/src/ATen/native/cpu/UnaryOpsKernel.cpp
@@ -258,6 +258,10 @@ void bernoulli_mkl_kernel(Tensor &output, const double p, Generator* gen) {
   // users to report this as a bug.
   AT_ASSERTM(false, "ATen not compiled with MKL");
 }
+
+void exponential_mkl_kernel(Tensor &output, const double p, Generator* gen) {
+  AT_ASSERTM(false, "ATen not compiled with MKL");
+}
 #else
 void bernoulli_mkl_kernel(Tensor &self, const double p, Generator* gen) {
   CPUGenerator* generator = get_generator_or_default<CPUGenerator>(gen, detail::getDefaultCPUGenerator());
@@ -306,6 +310,59 @@ void bernoulli_mkl_kernel(Tensor &self, const double p, Generator* gen) {
     // copy_ if using buffer and non contiguous
     if (!contig) {
       self.copy_(tmp_int_tensor);
+    }
+  });
+}
+
+static inline void vslExponential(const int method, VSLStreamStatePtr stream,
+    const int n, float* r, const float a, const float b) {
+  vsRngExponential(method, stream, n, r, a, b);
+}
+
+static inline void vslExponential(const int method, VSLStreamStatePtr stream,
+    const int n, double* r, const double a, const double b) {
+  vdRngExponential(method, stream, n, r, a, b);    
+}
+
+void exponential_mkl_kernel(Tensor &self, const double p, Generator* gen) {
+  CPUGenerator* generator = get_generator_or_default<CPUGenerator>(gen, detail::getDefaultCPUGenerator());
+  int64_t seed;
+  {
+    // See Note [Acquire lock when using random generators]
+    std::lock_guard<std::mutex> lock(generator->mutex_);
+    seed = generator->random();
+  }
+  int64_t n = self.numel();
+  bool contig = self.is_contiguous();
+
+  AT_DISPATCH_FLOATING_TYPES(self.scalar_type(), "exponential_cpu_", [&] {
+    at::Tensor tmp_tensor;
+    if (contig) {
+      tmp_tensor = self;
+    } else {
+      tmp_tensor = at::empty(self.sizes(), self.options());
+    }
+
+    scalar_t *self_ptr = self.data_ptr<scalar_t>();
+    scalar_t *sample_ptr = tmp_tensor.data_ptr<scalar_t>();
+
+    auto sample = [&](int64_t begin, int64_t end) {
+      int64_t len = end - begin;
+      if (len > 0) {
+        VSLStreamStatePtr stream;
+        vslNewStream(&stream, VSL_BRNG_MCG31, seed);
+        vslSkipAheadStream(stream, begin);
+        vslExponential(VSL_RNG_METHOD_EXPONENTIAL_ICDF, stream, len,
+          sample_ptr + begin, scalar_t(0), static_cast<scalar_t>(1. / p));
+        vslDeleteStream(&stream);
+      }
+    };
+
+    parallel_for(0, n, /* grain_size= */ 800, sample);
+
+    // copy_ if using buffer and non contiguous
+    if (!contig) {
+      self.copy_(tmp_tensor);
     }
   });
 }
@@ -391,6 +448,7 @@ static void rsqrt_kernel(TensorIterator& iter) {
 REGISTER_DISPATCH(rsqrt_stub, &rsqrt_kernel);
 REGISTER_DISPATCH(sigmoid_stub, &sigmoid_kernel);
 REGISTER_DISPATCH(bernoulli_mkl_stub, &bernoulli_mkl_kernel);
+REGISTER_DISPATCH(exponential_mkl_stub, &exponential_mkl_kernel);
 REGISTER_DISPATCH(abs_stub, &abs_kernel);
 REGISTER_DISPATCH(angle_stub, &angle_kernel);
 REGISTER_DISPATCH(real_stub, &real_kernel);