--- a/aten/src/TH/generic/THTensorRandom.h
+++ b/aten/src/TH/generic/THTensorRandom.h
@@ -17,7 +17,6 @@ TH_API void THTensor_(normal)(THTensor *self, double mean, double stdv, at::Gene
 TH_API void THTensor_(normal_means)(THTensor *self, THTensor *means, double stddev, at::Generator *gen);
 TH_API void THTensor_(normal_stddevs)(THTensor *self, double mean, THTensor *stddevs, at::Generator *gen);
 TH_API void THTensor_(normal_means_stddevs)(THTensor *self, THTensor *means, THTensor *stddevs, at::Generator *gen);
-TH_API void THTensor_(exponential)(THTensor *self, double lambda, at::Generator *_generator);
 TH_API void THTensor_(cauchy)(THTensor *self, double median, double sigma, at::Generator *_generator);
 TH_API void THTensor_(logNormal)(THTensor *self, double mean, double stdv, at::Generator *_generator);
 TH_API void THTensor_(multinomialAliasSetup)(THTensor *prob_dist, THLongTensor *J, THTensor *q);