@VitalyFedyunin , this PR is about move exponential distribution to Aten and also use MKL to improve the performance.
Benchmark:
```
import torch
import torch.nn as nn
import time

torch.manual_seed(0)

def _time():
    return time.time()

device = "cpu"

#warm up
for n in [10, 100, 1000]:
    input = torch.randn(128, n, requires_grad=False, device=device)
    for i in range(1000):
        input.exponential_()

for n in [1, 10, 100, 1000]:
    fwd_t = 0
    input = torch.randn(128, n, requires_grad=False, device=device)
    for i in range(10000):
        t1 = _time()
        input.exponential_()
        t2 = _time()
        fwd_t = fwd_t + (t2 -t1)
    fwd_avg = fwd_t / 10000 * 1000
    print("input size(128, %d) forward time is %.4f (ms)." % (n, fwd_avg))
```
Test device: **skx-8180**.
Before:
```
input size(128, 1) forward time is 0.0067 (ms).
input size(128, 10) forward time is 0.0560 (ms).
input size(128, 100) forward time is 0.5478 (ms).
input size(128, 1000) forward time is 5.4507 (ms).
```
After:
```
input size(128, 1) forward time is 0.0024 (ms).
input size(128, 10) forward time is 0.0053 (ms).
input size(128, 100) forward time is 0.0071 (ms).
input size(128, 1000) forward time is 0.0138 (ms).
```
Fix #24699