@XiaobingSuper 2019-12-27
Will add MKL path to improve the performance.