@XiaobingSuper 2019-12-30
@VitalyFedyunin , MKL path added to improve the performance, please see it.