@XiaobingSuper 2019-12-30
For Exponential distribution, lambda should be a positive numer, see https://en.wikipedia.org/wiki/Exponential_distribution, we should return an error as [numpy](https://docs.scipy.org/doc/numpy-1.9.0/reference/generated/numpy.random.exponential.html)'s doing when lambda is a negative number.